refactor: 폴더 구조 변경 및 footer 다크모드 반영
