+++ css/page/review.css
@@ -57,12 +57,12 @@
     display: block;
     width: 12px;
     height: 12px;
-    mask: url(../img/detail/star-fill.svg) no-repeat center /12px;
+    mask: url(../../img/detail/star-fill.svg) no-repeat center /12px;
     background-color: var(--yellow-color1);
 }
 
 #reviewContainer [class*='star-half'] {
-    mask-image: url(../img/detail/star-half.svg);
+    mask-image: url(../../img/detail/star-half.svg);
 }
 
 #reviewContainer .section-title {
