+++ index.html
@@ -18,7 +18,7 @@
       crossorigin
       href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css"
     />
-    <link rel="stylesheet" href="./css/footer.css" />
+    <link rel="stylesheet" href="css/page/footer.css" />
     <script src="./js/swiper-bundle.min.js"></script>
     <script defer src="./js/common.js" type="module"></script>
     <script defer src="./js/index.js" type="module"></script>
@@ -30,7 +30,7 @@
         <div class="header-nav">
           <nav class="header-nav-menu">
             <ul>
-              <li><a href="./profile.html">내 프로필</a></li>
+              <li><a href="./page/profile.html">내 프로필</a></li>
               <li><a class="not-open-link" href="#">이용권</a></li>
               <li><a class="not-open-link" href="#">쿠폰·코인</a></li>
               <li><a class="not-open-link" href="#">웨이브온</a></li>
@@ -40,9 +40,7 @@
           </nav>
         </div>
         <div class="header-section">
-          <a href="./index.html" class="logo"
-            ><span class="blind">Wavve</span></a
-          >
+          <a href="./index.html" class="logo"><span class="blind">Wavve</span></a>
           <nav class="header-menu">
             <ul>
               <li><a href="#" class="on">홈</a></li>
@@ -50,19 +48,19 @@
                 <a href="#">카테고리</a>
                 <div class="category">
                   <ul>
-                    <li><a href="./subpage/top20.html">TOP 20</a></li>
-                    <li><a href="./subpage/nowplaying.html">NOW PLAYING</a></li>
-                    <li><a href="./subpage/upcoming.html">UPCOMING</a></li>
-                    <li><a href="./subpage/popular.html">POPULAR</a></li>
+                    <li><a href="./page/top20.html">TOP 20</a></li>
+                    <li><a href="./page/nowplaying.html">NOW PLAYING</a></li>
+                    <li><a href="./page/upcoming.html">UPCOMING</a></li>
+                    <li><a href="./page/popular.html">POPULAR</a></li>
                   </ul>
                 </div>
               </li>
               <li><a class="not-open-link" href="#">LIVE</a></li>
-              <li><a href="./profile.html">MY</a></li>
+              <li><a href="./page/profile.html">MY</a></li>
             </ul>
           </nav>
           <div class="search-container">
-            <a href="./search.html">
+            <a href="./page/search.html">
               <img src="./img/common/icon-search.svg" alt="검색" />
             </a>
           </div>
@@ -90,7 +88,7 @@
       <section class="cards-section top-rated">
         <div class="cards-section-title">
           TOP 20
-          <a href="./subpage/top20.html" class="more">
+          <a href="./page/top20.html" class="more">
             더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
           </a>
         </div>
@@ -108,7 +106,7 @@
       <section class="cards-section playing">
         <div class="cards-section-title">
           NOW PLAYING
-          <a href="./subpage/nowplaying.html" class="more">
+          <a href="./page/nowplaying.html" class="more">
             더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
           </a>
         </div>
@@ -126,7 +124,7 @@
       <section class="cards-section upcoming">
         <div class="cards-section-title">
           UPCOMING
-          <a href="./subpage/upcoming.html" class="more">
+          <a href="./page/upcoming.html" class="more">
             더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
           </a>
         </div>
@@ -144,7 +142,7 @@
       <section class="cards-section popular">
         <div class="cards-section-title">
           POPULAR
-          <a href="./subpage/popular.html" class="more">
+          <a href="./page/popular.html" class="more">
             더보기<img src="./img/main/arrow-more.svg" class="arrow-more" />
           </a>
         </div>
@@ -218,10 +216,8 @@
       <div class="footer-menu">
         <div class="footer-wavveon">
           <div class="logo">
-            <img
-              src="https://www.wavve.com/img/ico-footer-wavve-on.d4912b54.svg"
-              alt="Wavve on"
-            />
+            <span class="logo-img"></span>
+            <span class="blind">wavve on</span>
           </div>
           <select>
             <option value disabled selected hidden>사업자용 웨이브 가입</option>
