+++ js/page/review/reviewForm/starRatingHandler.js
