+++ page/detail.html
@@ -4,15 +4,15 @@
     <meta charset='UTF-8'>
     <title>Energivve(에너자이브)</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no"/>
-    <link rel="shortcut icon" href="./img/favicon.ico"/>
-    <link rel="stylesheet" href="./css/reset.css"/>
-    <link rel="stylesheet" href="./css/common.css"/>
-    <link rel="stylesheet" href="./css/detail.css"/>
-    <link rel="stylesheet" href="./css/review.css"/>
+    <link rel="shortcut icon" href="../img/favicon.ico"/>
+    <link rel="stylesheet" href="../css/reset.css"/>
+    <link rel="stylesheet" href="../css/common.css"/>
+    <link rel="stylesheet" href="../css/page/detail.css"/>
+    <link rel="stylesheet" href="../css/page/review.css"/>
     <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css" />
-    <script defer src="./js/common.js" type="module"></script>
-    <script defer src="./js/detail.js" type="module"></script>
-    <script defer src="js/review/index.js" type="module"></script>
+    <script defer src="../js/common.js" type="module"></script>
+    <script defer src="../js/page/detail.js" type="module"></script>
+    <script defer src="../js/page/review/index.js" type="module"></script>
 </head>
 <div id="loaderWrap"><div id="loader"></div></div>
 <body id="detailPage">
@@ -31,28 +31,28 @@
             </nav>
         </div>
         <div class="header-section">
-            <a href="./index.html" class="logo"><span class="blind">Wavve</span></a>
+            <a href="../index.html" class="logo"><span class="blind">Wavve</span></a>
             <nav class="header-menu">
                 <ul>
-                    <li><a href="./index.html" class="on">홈</a></li>
+                    <li><a href="../index.html" class="on">홈</a></li>
                     <li>
                         <a href="#">카테고리</a>
                         <div class="category">
                             <ul>
-                                <li><a href="./subpage/top20.html">TOP 20</a></li>
-                                <li><a href="./subpage/nowplaying.html">NOW PLAYING</a></li>
-                                <li><a href="./subpage/upcoming.html">UPCOMING</a></li>
-                                <li><a href="./subpage/popular.html">POPULAR</a></li>
+                                <li><a href="./top20.html">TOP 20</a></li>
+                                <li><a href="./nowplaying.html">NOW PLAYING</a></li>
+                                <li><a href="./upcoming.html">UPCOMING</a></li>
+                                <li><a href="./popular.html">POPULAR</a></li>
                             </ul>
                         </div>
                     </li>
                     <li><a class="not-open-link" href="#">LIVE</a></li>
-                    <li><a href="./profile.html">MY</a></li>
+                    <li><a href="profile.html">MY</a></li>
                 </ul>
             </nav>
             <div class="search-container">
-                <a href="./search.html">
-                    <img src="./img/common/icon-search.svg" alt="검색" />
+                <a href="search.html">
+                    <img src="../img/common/icon-search.svg" alt="검색" />
                 </a>
             </div>
         </div>
