+++ page/profile.html
@@ -4,17 +4,17 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no"/>
     <title>프로필 설정 : 에너자이브</title>
-    <link rel="shortcut icon" href="./img/favicon.ico"/>
-    <link rel="stylesheet" href="./css/reset.css"/>
-    <link rel="stylesheet" href="./css/common.css"/>
-    <link rel="stylesheet" href="./css/profile.css"/>
-    <script defer src="./js/common.js" type="module"></script>
-    <script defer src="./js/profile.js"></script>
+    <link rel="shortcut icon" href="../img/favicon.ico"/>
+    <link rel="stylesheet" href="../css/reset.css"/>
+    <link rel="stylesheet" href="../css/common.css"/>
+    <link rel="stylesheet" href="../css/page/profile.css"/>
+    <script defer src="../js/common.js" type="module"></script>
+    <script defer src="../js/page/profile.js"></script>
   </head>
   <body>
     <div class="container">
       <div class="logo_section">
-        <a href="index.html">
+        <a href="../index.html">
           <span class="blind">Wavve</span>
         </a>
       </div>
@@ -56,7 +56,7 @@
           />
         </div>
         <div class="confirm_box">
-          <a href="index.html" id="cancel" class="btn cancel">취소</a>
+          <a href="../index.html" id="cancel" class="btn cancel">취소</a>
           <button id="confirm" type="submit" class="btn confirm">확인</button>
         </div>
       </form>
