+++ page/search-result.html
@@ -4,14 +4,14 @@
     <meta charset="UTF-8" />
     <title>검색 결과 : Energivve(에너자이브)</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no"/>
-    <link rel="shortcut icon" href="./img/favicon.ico"/>
-    <link rel="stylesheet" href="./css/reset.css"/>
-    <link rel="stylesheet" href="./css/common.css"/>
-    <link rel="stylesheet" href="./css/search.css"/>
-    <link rel="stylesheet" href="./css/search-result.css" />
+    <link rel="shortcut icon" href="../img/favicon.ico"/>
+    <link rel="stylesheet" href="../css/reset.css"/>
+    <link rel="stylesheet" href="../css/common.css"/>
+    <link rel="stylesheet" href="../css/page/search.css"/>
+    <link rel="stylesheet" href="../css/page/search-result.css" />
     <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css"/>
-    <script defer src="./js/common.js" type="module"></script>
-    <script defer src="./js/searchList.js" type="module"></script>
+    <script defer src="../js/common.js" type="module"></script>
+    <script defer src="../js/page/searchList.js" type="module"></script>
 </head>
 <body>
 <header>
@@ -29,18 +29,18 @@
             </nav>
         </div>
         <div class="header-section">
-            <a href="./index.html" class="logo"><span class="blind">Wavve</span></a>
+            <a href="../index.html" class="logo"><span class="blind">Wavve</span></a>
             <nav class="header-menu">
                 <ul>
-                    <li><a href="./index.html" class="on">홈</a></li>
+                    <li><a href="../index.html" class="on">홈</a></li>
                     <li>
                         <a href="#">카테고리</a>
                         <div class="category">
                             <ul>
-                                <li><a href="./subpage/top20.html">TOP 20</a></li>
-                                <li><a href="./subpage/nowplaying.html">NOW PLAYING</a></li>
-                                <li><a href="./subpage/upcoming.html">UPCOMING</a></li>
-                                <li><a href="./subpage/popular.html">POPULAR</a></li>
+                                <li><a href="./top20.html">TOP 20</a></li>
+                                <li><a href="./nowplaying.html">NOW PLAYING</a></li>
+                                <li><a href="./upcoming.html">UPCOMING</a></li>
+                                <li><a href="./popular.html">POPULAR</a></li>
                             </ul>
                         </div>
                     </li>
@@ -49,10 +49,10 @@
                 </ul>
             </nav>
             <form id="searchContainer">
-                <img src="img/common/icon-search.svg" class="search-input-icon" />
+                <img src="../img/common/icon-search.svg" class="search-input-icon" />
                 <input id="searchInput" placeholder="제목, 장르, 배우로 찾아보세요" type="text" class="search-input-box"/>
                 <button id="searchDeleteBtn" type="button">
-                    <img src="./img/common/search-delete.svg" class="search-cancel-icon" />
+                    <img src="../img/common/search-delete.svg" class="search-cancel-icon" />
                 </button>
             </form>
         </div>
