+++ page/top20.html
@@ -3,22 +3,22 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0, user-scalable=no"/>
-    <title>upcoming : Energivve(에너자이브)</title>
+    <title>top20 : Energivve(에너자이브)</title>
     <link rel="shortcut icon" href="../img/favicon.ico" />
     <link rel="stylesheet" href="../css/reset.css" />
     <link rel="stylesheet" href="../css/common.css" />
-    <link rel="stylesheet" href="../css/subpage.css" />
+    <link rel="stylesheet" href="../css/page/subpage.css" />
     <link rel="stylesheet" as="style" crossorigin href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard@v1.3.9/dist/web/static/pretendard-dynamic-subset.min.css"/>
     <script defer src="../js/common.js" type="module"></script>
-    <script defer src="../js/subpage.js" type="module"></script>
+    <script defer src="../js/page/subpage.js" type="module"></script>
   </head>
-  <body id="upcomingPage">
+  <body id="topPage">
     <header>
       <div class="header-wrap">
         <div class="header-nav">
           <nav class="header-nav-menu">
             <ul>
-              <li><a href="../profile.html">내 프로필</a></li>
+              <li><a href="./profile.html">내 프로필</a></li>
               <li><a class="not-open-link" href="#">이용권</a></li>
               <li><a class="not-open-link" href="#">쿠폰·코인</a></li>
               <li><a class="not-open-link" href="#">웨이브온</a></li>
@@ -36,19 +36,19 @@
                 <a href="#">카테고리</a>
                 <div class="category">
                   <ul>
-                    <li><a href="../subpage/top20.html">TOP 20</a></li>
-                    <li><a href="../subpage/nowplaying.html">NOW PLAYING</a></li>
-                    <li><a href="../subpage/upcoming.html">UPCOMING</a></li>
-                    <li><a href="../subpage/popular.html">POPULAR</a></li>
+                    <li><a href="./top20.html">TOP 20</a></li>
+                    <li><a href="./nowplaying.html">NOW PLAYING</a></li>
+                    <li><a href="./upcoming.html">UPCOMING</a></li>
+                    <li><a href="./popular.html">POPULAR</a></li>
                   </ul>
                 </div>
               </li>
               <li><a class="not-open-link" href="#">LIVE</a></li>
-              <li><a href="../profile.html">MY</a></li>
+              <li><a href="profile.html">MY</a></li>
             </ul>
           </nav>
           <div class="search-container">
-            <a href="../search.html">
+            <a href="search.html">
               <img src="../img/common/icon-search.svg" alt="검색" />
             </a>
           </div>
@@ -57,9 +57,9 @@
     </header>
     <div class="page-top">
       <div class="page-top-inner-wrap">
-        <h1 class="page-title">Upcoming</h1>
-        <p class="page-text">곧 만나게 될</p>
-        <p class="page-text">기대되는 영화 라인업을 알아보세요</p>
+        <h1 class="page-title">TOP 20</h1>
+        <p class="page-text">오랫동안 사랑받는</p>
+        <p class="page-text">인기 있는 TOP 20위 영화를 확인하세요</p>
       </div>
     </div>
     <ul class="cards-section"></ul>
